requirements.txt
# LangChain and LLM
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langchain-chroma>=0.1.0

# Vector Store
chromadb>=0.5.0

# Data validation
pydantic>=2.0.0

# Template engines
jinja2>=3.1.0
python-liquid>=1.12.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Utilities
python-dotenv>=1.0.0
